requirements.txt
asgiref==3.8.1
Django==3.1.12
djangorestframework==3.12.4
djongo==1.3.7
dnspython==2.7.0
pymongo==3.11.4
pytz==2025.1
sqlparse==0.2.4
typing_extensions==4.12.2
